requirements.txt
Flask
requests
beautifulsoup4
lxml
feedparser
vaderSentiment
SQLAlchemy
pymysql         # MySQL driver (if using MySQL)
python-dotenv
